requirements.txt
# Flask and Extensions
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.4.4
Flask-Cors==3.0.10
Flask-Migrate==4.0.4

# Celery and Redis for Task Scheduling
celery==5.2.7
redis==4.6.0

# Database and ORM
psycopg2-binary==2.9.7
SQLAlchemy==2.0.16

# Validation and Security
Werkzeug==2.3.7

# Environment Configuration
python-dotenv==1.0.0

# API Documentation
flask-swagger-ui==4.10.1

# Logging and Monitoring
gunicorn==20.1.0

# AWS SDK
boto3==1.28.27

# AI Integration (Optional, based on OpenAI usage)
openai==0.27.8

# Testing Tools
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.10.0

# Utilities
uuid==1.30
validators==0.20.0

# Asynchronous Task Management
flower==1.2.0

# Optional: Additional Libraries for Enhancements
requests==2.31.0
